refactor: renamed routes to pages
