+++ lib/src/pages/edit_currency_state.dart
